feat: responsive search bar placement
